video_core: refactored post processing pipeline to another file
