+++ src/video_core/renderer_vulkan/host_passes/pp_pass.h
@@ -1,0 +1,34 @@
+//  SPDX-FileCopyrightText: Copyright 2024 shadPS4 Emulator Project
+//  SPDX-License-Identifier: GPL-2.0-or-later
+
+#pragma once
+
+#include "common/types.h"
+#include "video_core/renderer_vulkan/vk_common.h"
+
+namespace Vulkan {
+class Frame;
+}
+
+namespace Vulkan::HostPasses {
+
+class PostProcessingPass {
+public:
+    struct Settings {
+        float gamma = 1.0f;
+        u32 hdr = 0;
+    };
+
+    void Create(vk::Device device);
+
+    void Render(vk::CommandBuffer cmdbuf, vk::ImageView input, vk::Extent2D input_size,
+                Frame& output, Settings settings);
+
+private:
+    vk::UniquePipeline pipeline{};
+    vk::UniquePipelineLayout pipeline_layout{};
+    vk::UniqueDescriptorSetLayout desc_set_layout{};
+    vk::UniqueSampler sampler{};
+};
+
+} // namespace Vulkan::HostPasses
